Rename FunctionInstanceType to FunctionType
