+++ PHP/Types/TypeLookup.php
@@ -4,7 +4,6 @@ declare( strict_types = 1 );
 namespace PHP\Types;
 
 use PHP\Types\Models\FunctionType;
-use PHP\Types\Models\FunctionInstanceType;
 use PHP\Types\Models\Type;
 use PHP\Types\TypeNames;
 
@@ -195,13 +194,13 @@ class TypeLookup
 
 
     /**
-     * Create a FunctionInstance type instance
+     * Create a FunctionType instance
      * 
      * @param \ReflectionFunction $function The ReflectionFunction instance for the function instance
      * @return Type
      */
-    protected function createFunctionInstanceType( \ReflectionFunction $function ): FunctionInstanceType
+    protected function createFunctionInstanceType( \ReflectionFunction $function ): FunctionType
     {
-        return new FunctionInstanceType( $function );
+        return new FunctionType( $function );
     }
 }
